+++ lib/enums/line_types.dart
