+++ lib/pages/search_page.dart
@@ -1,7 +1,7 @@
-import 'package:departures/line_types.dart';
+import 'package:departures/enums/line_types.dart';
 import 'package:departures/services/stop.dart';
-import 'package:departures/services/vbb_api.dart';
-import 'package:departures/station_display.dart';
+import 'package:departures/services/api.dart';
+import 'package:departures/widgets/station_display.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
