+++ lib/provider/theme_settings_provider.dart
@@ -1,23 +1,23 @@
-import 'package:departures/provider/theme_modes.dart';
+import 'package:departures/enums/app_theme_modes.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class ThemeProvider with ChangeNotifier {
-  AppThemeMode _themeModeEnum = AppThemeMode.system;
-  AppThemeMode get themeModeEnum => _themeModeEnum;
+  AppThemeMode _appThemeMode = AppThemeMode.system;
+  AppThemeMode get appThemeMode => _appThemeMode;
 
-  bool get materialYou => _themeModeEnum == AppThemeMode.you;
+  bool get materialYou => _appThemeMode == AppThemeMode.you;
 
   ThemeProvider() {
     loadThemePreference();
   }
 
   Future<void> saveThemePreference(AppThemeMode themeMode) async {
-    _themeModeEnum = themeMode;
+    _appThemeMode = themeMode;
     notifyListeners();
 
     SharedPreferences prefs = await SharedPreferences.getInstance();
-    prefs.setInt('themeMode', _themeModeEnum.index);
+    prefs.setInt('themeMode', _appThemeMode.index);
   }
 
   Future<void> loadThemePreference() async {
@@ -25,7 +25,7 @@ class ThemeProvider with ChangeNotifier {
     int? themeIndex = prefs.getInt('themeMode');
 
     if (themeIndex != null) {
-      _themeModeEnum = AppThemeMode.values[themeIndex];
+      _appThemeMode = AppThemeMode.values[themeIndex];
       notifyListeners();
     }
   }
